Lendsqr API Assessment FIle
How to Run the Script:

Step 1. Either do a pull request or do a direct download of the attachment (Lendsqr Assessment.postman_collection.json) to your local machine.
Step 2. Open your Postman application either on the cloud or your desktop.
Step 3. Create a workspace on the Postman application.
Step 4. Create a Collection inside the workspace.
Step 5. Click on the import button at the left top corner of the postman.
Step 6. Drag the downloaded file in Step 1 into the Postman upload interface.
Step 7. Waits for Postman to complete the upload.
Step 8. Expand the imported collection and the modules to access the HTTP requests.
Step 9. Click on each of the HTTP requests to open it up on Postman.
Step 10. Click on the Scripts tab to assess the test written for each request.
Step 11. Click on the Send button to initiate each request and compare results.